Lockscreen charging info: show decimal in battery temperature

+++ packages/SystemUI/src/com/android/systemui/statusbar/KeyguardIndicationController.java
@@ -197,7 +197,7 @@ public class KeyguardIndicationController {
     private final FaceHelpMessageDeferral mFaceAcquiredMessageDeferral;
     private int mChargingCurrent;
     private double mChargingVoltage;
-    private int mTemperature;
+    private float mTemperature;
     private String mMessageToShowOnScreenOn;
     private boolean mInited;
